requirements.txt
pandas == 2.2.2
PyYAML == 6.0.1
folium == 0.17.0
jupyterlab == 4.2.1